requirements.txt
numpy
pillow
matplotlib
tqdm
scikit-image
scikit-learn
pandas
opencv-python
torch
torchvision
torchaudio
kaggle
pennylane
pennylane-lightning
